Fixies Rails bug #2521 - auto dasherizing. Now defaults to "do nothing"
Currently ActiveResource forces all xml to be dasherised even if we don't
pass in :dasherzie => true.
There's no way to override this anywhere, and the "change the values"
default fails the principle of least surprise.

This is an issue when integrating with any system that is not rails...

This bugfix changes the default behaviour to default to "do nothing" ie you
can pass in :dasherize => true if that's what you want, but if you don't
pass that in, then the defalt behaviour is to keep it in underscore case.