bugfix: validates_uniqueness_of was only working on create, not update...
So now I've put an experimental fix in to make it check on update that the
record that matches doesn't have the same id as us.

I've also added the "allow_nil" config option which skips this check if the
value is *blank* (because IMO that's more useful than nil).

I've also changed the find to pass the attr in the conditions block so that
if there's a Rails API behind this (that honours a conditions block) it will
only pass back the records that match... potentially reducing traffic.