Better handling of 404s in destroy_all/delete_all
There's no reason why destroy_all should throw a 404 - we can just swallow
that nicely like ActiveRecord does when nothing is found.
If you get a 404 - you can be certain that everything was destroyed!
In this case I actually return a nil (instead of true) so you can even test
for when that occurred.