Attribute-massaging moved to encode method to allow multiple formats.
Updated the code in to_xml to remove all the special attribute-massaging
(for nested attribute ids) and to put it into the more-generic "encode"
method (so it can be used for to_json etc as well).

All the update functions here now call encode (which is used as standard in
ActiveResource) so that HyRes is not artificially restricted to xml-only
applications.

I've also added a quick short-circuit when there are no associations to
bother with - so we don't need to duplicate the attibutes if unnecessary...

Downside: I've had to duplicate the formatter-encoding when we use massaged
attributes - this will need to be kept up-to-date with Rails' way...
It might be a better idea to overload the "attributes" method to always
return the massaged attributes and then this wouldn't be necessary?