Extended collection_path to allow rails named routes and better count uses it
Extended collection_path to allow us to pass in suffix_options that will be
put after collection_name and before the format-extension - thus allowing
standard rails-style named routes.

Also updated the .count function to use this and try a count_widgets path
equivalent rather than fetching *all* items and then counting (the slow
way). More doco on usage of that is on the method itself, but in breif, you
can use the default railsish-style count_widgets path, or pass in a
per-model-basis path or override all fo the above any time you ask for a count.