Better functioning of the find_by_X methods
1) now combines conditions (rather than assuming that the chosen field is
the only argument).
2) updated find_every to add the "params" key if it's not been explicitly
passed in. That allows us to write ActiveRecord-style finders without having
to worry about an extraneous "params" key. Note: it boviously doesn't pass
it in if it already exists and it also preserves a "from" key if it exists -
so code should still work as before.
3) added a rudimentary "merge_conditions" method to allow us to merge in
conditions.
4) updated the README to include conditions passed into finders.