added the validations explicitly to create/update so we don't bother hitting the remote network if the local object isn't valid.
I had been relying on doing validations via save - but it seems that
sometimes we don't get there that way... and I wanted to be sure
validations are performed.

This may be removed later when we can be sure all vectors into create/update
go via save - but for now it makes sure.

Updated the collection_path to pull suffix_options out of the generic set of
options (if they haven't been passed-in specifically). This allows us to
pass them in via conditional finders which is a more complex version of
setting up a from-path than simply having to pass in a pre-grenerated
string. This will allow us to generate more complicated paths on the fly -
eg by other HyRes methods.

Updated find_every - now no longer rely on ActiveResources' version and just
reproduce the functionality here with our own changes baked in. This is so
that we can use our own "collection_path" with the optional suffix_options -
and thus implement calculate paths (eg Widget.count) with optional
conditions.

Some spell-fixes

Allow an instance to create a collection_apth by calling the same options on
the parent class.