Updated has_many fetcher - actually works!
at leats i think so.
Before it only fetched out existing ids if they'd already been assigned to
the object - and didn't go back to the associated class unless we already
had ids. Now we do a find on the associated object passing in the self.id to
fetch all the associated objects out in one set.