Fixed find so it doesn't 404 where it shouold be returning nil
In ActiveRecord - when a collection-finder can't find anything - it doesn't
return a 404-equivalent, it just returns nil.
This is especially the case for associations eg @user.sites shouldn't 404 -
it should just return nil.

So this is what I've done. For all finders apart from the finds by id (eg
Thingy.find(1)) we now return nil instead of raising an exception.