Add an option to belongs_to that allows for nested routes on the remote API.
If your remote API is Rails-ish with nested resource routes - we want to be
able to set up our Hyres objects without too much hassle. Firtly - it's
annoynig to have to type in the prefix-path when it's rails-standard, and
secondly - it's hard to pass in the id-values if we're inside the
association's collection-fetch function.

So - I've added a "nested" option for the belongs_to relationship which will
automatically do all these things for me.