paramsify the options for find(<id>)
Just started using find_single instead of find_every and realised I never
put a paramsification of the options on this one.
So now it can be used a la AR:
Thingy.find(4, :condtions => {:user_id => 3})

why do this? well, if you want to fetch an object - but only if the
current_user owns the object... you need to pass conditions to the backend.