Finally got a reasonably-working split_options.
A lot of mucking-about to get to this point. Most of the previous checkins
are non-working to some extent.
The issue began with nested URLs when you pass through the nested id as a
condition eg for a class Wdiegt with a nested URL such as:
/users/:user_id/widgets
Calling:
Widget.find(:first, :conditions => {:user_id => 42})
would create a URL:
/users//widgets?conditions[user_id]=42

which would break due to a lack of :user_id... even though it's present in
the conditions.

So I had to update the split_options method to search the conditions-string
for possible prefix_parameters.

Then it was passing it in both eg:

/users/42/widgets?conditions[user_id]=42

Which is a bit icky doing double-filtering and sometimes broke the remote
API stuff, so I updated split_options to remove any query_options that were
already present in the prefix_options.

For some reason one of these changes started introducing a whole bunch of
null options - because the prefix_options start out with empty values (eg
:select => nil), so I had to remove any keys that had a null value...
but that broke the load method which opulls out attributes... which have
null values that are meaningful... so I had to create a new option that can
be passed in which allows me to keep null values if thay are meaningful.

Exceptions to exceptions... :P

But it now seems to cover all bases and not break previously-working code...