Bugfix to from_xml when an instance passed instead of xml
when from_xml-ing, if we haven't actually passed in xml - but have passed an
instance of ourselves... we don't need to from_xml ourselves.

This might happen on reload of an instance that has already from_xmled some
associations.

Also - a record should be able to instantiate from an instance of itself by
duplicating!