to_params-ification done so primary_key = :uuid should work now
There were a number of references directly to ".id" in the code. These have
all been replaced by "to_param" so we can use the pre-existing ARes
primary_key function.
The "id_from_response" was also failing because it didn't reference the
primary_key, but "id".

Also updated README to better comment the example and give an idea of a
callback to generate a uuid...