Fix for latest stable version of rails' validations
The current version of "generate_message" does something weird to collect up
the superclasses of active record - using a method called self_and_descendants_from_active_record
obviously this is named inappropriately for general piracy and use
within HyRes... but no choice but to cover it for the meantime.

This fix will work even if it's not pretty and largely irrelevant to how
HyRes works atm.