Update to the raw_data to ensure format is reset back.
The bug was that if we make a request on an object in a new format - we set
the format to, say, 'pdf'. Then, if we get a 404 or 500 etc on the call for
this pdf - we could escape without resetting the format back to the
original (eg xml)... which means every call after then will try (and fail)
to decode from pdf...

This change will ensure that we reset the format back to the original
format regardless of the success of the get.