Count-method fix: respond better to "find with id = 'count'" response
A RESTful interface may easily accept the path: /widgets/count.xml
and turn it into:
  find widgets with the id = "count"
  oops, can't find one... return

If it returns a 404, we already catch this... this is the equivalent to a
find_single or find_by_id request... but some interfaces will convert the
above into a 'find all matching' request, where the correct response is a
null-object.

So we don't get an exception, just an object that doesn't respond to the
'count' method (eg an empty array).

This fix - checks for that.

(phew, long explanation for a simple fix!) :)